+++ lib/online_migrations/background_schema_migrations/migration_runner.rb
@@ -44,9 +44,8 @@ module OnlineMigrations
         # Background schema migrations could take a while to run. It is possible, that the process
         # never reaches this (or the rescue below) line of code. E.g., when it is force quitted
         # (SIGKILL etc.) and so the migration will end up in the "running" state and the query is
-        # still executing (or already finished) in the database. This migration can either be safely
-        # manually retried or will be picked up in the future by scheduler when it decides that
-        # this migration is stuck.
+        # still executing (or already finished) in the database. This migration can be safely
+        # manually retried.
 
         migration.update!(status: :succeeded, finished_at: Time.current)